seperate encode/decode from main script
