+++ wrapper.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from bitshuffle import main
+main()
